+++ src/example/pegtl/dynamic_match.cpp
@@ -61,7 +61,6 @@ namespace dynamic
 
    template< typename Rule >
    struct action
-      : pegtl::nothing< Rule >
    {
    };