+++ src/example/pegtl/hello_world.cpp
@@ -18,7 +18,6 @@ namespace hello
 
    template< typename Rule >
    struct action
-      : pegtl::nothing< Rule >
    {
    };