+++ src/example/pegtl/indent_aware.cpp
@@ -101,7 +101,6 @@ namespace example
 
    template< typename Rule >
    struct action
-      : pegtl::nothing< Rule >
    {
    };