+++ src/example/pegtl/symbol_table.cpp
@@ -43,7 +43,6 @@ namespace example
 
    template< typename Rule >
    struct action
-      : pegtl::nothing< Rule >
    {
    };