+++ src/test/pegtl/rule_if_apply.cpp
@@ -65,7 +65,7 @@ namespace tao
          };
 
          template< typename Rule >
-         struct action : nothing< Rule >
+         struct action
          {
          };