+++ src/test/pegtl/rule_opt.cpp
@@ -10,7 +10,7 @@ namespace tao
    namespace TAO_PEGTL_NAMESPACE
    {
       template< typename Rule >
-      struct my_action : nothing< Rule >
+      struct my_action
       {
       };