+++ src/test/pegtl/rule_until.cpp
@@ -26,7 +26,7 @@ namespace tao
       };
 
       template< typename Rule >
-      struct my_action : nothing< Rule >
+      struct my_action
       {
       };